+++ packages/ui-core/src/components/Accordion/Accordion.docz.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+export const DemoParagraphWrapper = ({ children }) => {
+    return (
+        <div style={{ padding: '24px 0' }}>
+            {children}
+        </div>
+    );
+};